rag_search/requirements.txt
# RAG Search 實驗環境依賴
google-genai>=1.0.0
python-dotenv>=1.0.0
